Added limits to the Admin
Basically, before when the Admin goes into my private pictures, it was the same page as if I was going to log in and into my private pics. It had the same functions, such as uploading/deleting pics. Also, my page could add points to my score, which was a problem. This fixed all of these problems